% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesNonBiasCorrected.R
\name{BayesNonBiasCorrected}
\alias{BayesNonBiasCorrected}
\title{Fit non-bias corrected model with missing outcomes in multivariate meta-analysis (MMA)}
\usage{
BayesNonBiasCorrected(y1, s1, y2, s2, seed=NULL, burn=10000, nmc=50000)
}
\arguments{
\item{y1}{The first outcome in our MMA. Missing outcomes should be denoted by NA.}

\item{s1}{Standard errors (SEs) for y1. The missing SEs should be denoted by NA.}

\item{y2}{The second outcome in our MMA. Missing outcomes should be denoted by NA.}

\item{s2}{Standard errors (SEs) for y2. The missing SEs should be denoted by NA.}

\item{seed}{Optional seed. A seed has to be specified in order to reproduce the exact results on a data set.}

\item{burn}{Number of burn-in iterations for MCMC. The default is 10,000.}

\item{nmc}{Number of total MCMC iterations. The default is 50,000.}

}
\value{
A list of the posterior point estimates ([parameter.hat]) and posterior samples ([parameter].samples) for the model parameters. For example, mu1.hat and mu2.hat give the posterior means of the population treatment effects, and tau1.hat and tau2.hat give the posterior means for the between-study heterogeneity parameters. All the MCMC samples (e.g. mu1.samples and mu2.samples) are also returned, so the user can obtain posterior credible intervals and other summary statistics or plot the posterior densities.
}
\description{
This function accepts the first outcome (y1), the standard errors (SEs) of the first outcome (s1), the second outcome (y2), and the SEs for the second outcome (s2). The function returns posterior point estimates and posterior samples for the parameters under the non-bias corrected model.
}
\examples{
# Real data from an MMA on the effects of exercise therapy on lower back pain
y1 <- c(-25.30, -1.70, 0.00, -4.90, -1.80, -1.25, -1.72, -3.90, -4.50,
        -3.00, -0.70, -1.09, -10.87, NA, NA, NA)
s1 <- c(6.5464359, 3.3338820, 0.5545626, 2.9509468, 1.1016252, 3.7911903,
        4.4478104, 1.6160652, 3.3036029, 2.4320922, 2.2312296, 1.0242034,
        6.3042946, NA, NA, NA)
y2 <- c(-29.00, -5.00, -6.10, -10.00, NA, -1.00, -8.10, -11.100, -12.00,
        -4.00, -1.80, -3.75, 5.00, 0.00, 5.00, -12.00)
s2 <- c(6.103570, 3.705613, 3.649380, 3.776989, NA, 3.687818, 7.573094,
        3.307464, 3.646706, 7.918978, 4.551576, 4.396285, 5.817653, 1.189535,
        4.136892, 3.974209)
study_sizes <- c(75, 122, 125, 71, 286, 50, 54, 196, 50, 47, 138,
                 49, 110, 115, 69, 57)

# Fit non-bias corrected model
nobias.mod <- BayesNonBiasCorrected(y1, s1, y2, s2, seed=12345,
                                    burn=10000, nmc=50000)

# Point estimates for (mu1, mu2, tau1, tau2) in non-bias corrected model
nobias.mod$mu1.hat
nobias.mod$mu2.hat
nobias.mod$tau1.hat
nobias.mod$tau2.hat

# 95 percent credible intervals for mu1 and mu2
mu1.nobiascor.CI <- quantile(nobias.mod$mu1.samples, probs=c(0.025,0.975))
mu2.nobiascor.CI <- quantile(nobias.mod$mu2.samples, probs=c(0.025,0.975))

mu1.nobiascor.CI
mu2.nobiascor.CI
}
