% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_ORB.R
\name{D_ORB}
\alias{D_ORB}
\title{Compute the D measure for quantifying the impact of outcome reporting bias (ORB)}
\usage{
D_ORB(samples.ABSORBmodel, samples.nobiasmodel)
}
\arguments{
\item{samples.ABSORBmodel}{MCMC samples under ABSORB model}

\item{samples.nobiasmodel}{MCMC samples under non-bias corrected model}

}
\value{
The D measure for quantifying the impact of ORB on the MMA)
}
\description{
This function accepts the MCMC samples from an ABSORB model fit and a non-bias corrected model fit. The function then computes the D measure between these two posterior densities. D measures lie between 0 and 1, with values close to 0 indicating more negligible impact from ORB and values close to 1 indicating more severe impact from ORB.
}
\examples{
# Real data from an MMA on the effects of exercise therapy on lower back pain
y1 <- c(-25.30, -1.70, 0.00, -4.90, -1.80, -1.25, -1.72, -3.90, -4.50,
        -3.00, -0.70, -1.09, -10.87, NA, NA, NA)
s1 <- c(6.5464359, 3.3338820, 0.5545626, 2.9509468, 1.1016252, 3.7911903,
        4.4478104, 1.6160652, 3.3036029, 2.4320922, 2.2312296, 1.0242034,
        6.3042946, NA, NA, NA)
y2 <- c(-29.00, -5.00, -6.10, -10.00, NA, -1.00, -8.10, -11.100, -12.00,
        -4.00, -1.80, -3.75, 5.00, 0.00, 5.00, -12.00)
s2 <- c(6.103570, 3.705613, 3.649380, 3.776989, NA, 3.687818, 7.573094,
        3.307464, 3.646706, 7.918978, 4.551576, 4.396285, 5.817653, 1.189535,
        4.136892, 3.974209)
study_sizes <- c(75, 122, 125, 71, 286, 50, 54, 196, 50, 47, 138,
                 49, 110, 115, 69, 57)

# Fit ABSORB model
absorb.mod <- ABSORB(study_sizes, y1, s1, y2, s2, seed=12345,
                     burn=10000, nmc=50000)

# Fit non-bias corrected model
nobias.mod <- BayesNonBiasCorrected(y1, s1, y2, s2, seed=12345,
                                    burn=10000, nmc=50000)

# Compute D measure for first outcome
D.mu1 <- D_ORB(absorb.mod$mu1.samples, nobias.mod$mu1.samples)

# Compute D measure for second outcome
D.mu2 <- D_ORB(absorb.mod$mu2.samples, nobias.mod$mu2.samples)

# Compute D measure for first and second endpoints jointly
D.mu1mu2 <- D_ORB(cbind(absorb.mod$mu1.samples, absorb.mod$mu2.samples),
                  cbind(nobias.mod$mu1.samples, nobias.mod$mu2.samples))

D.mu1
D.mu2
D.mu1mu2
}
